Moving the container and presenter into `DatastoreNavigation`.
